str_buf_free_buf: Add null check (#37)
* os_timer_sig: rename "restart" to "restart_with_period", add "relaunch""

* str_buf_free_buf: Add null check

A new check has been added to the str_buf_free_buf method to handle null pointers. This ensures that os_free is not called with a null pointer which could potentially cause errors. The change includes a new test case in test_str_buf.cpp to verify handling of null pointers.

* Add esp_log_level_set to tests/.../esp_log.h